10_LLM-service/02_langchain/01_lecture-source/04_simple-chatbot.py/02-01_simple-chatbot-serving/requirements.txt
fastapi
uvicorn
gunicorn
langchain
langchain-community
langchain-openai
python-dotenv
python-multipart